Correcting int/long mismatched - 64bit issues.
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@119751 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/standard/url_scanner.c
@@ -88,7 +88,7 @@ static char *url_attr_addon(const char *tag,const char *attr,const char *val,con
 
 /* {{{ url_adapt_ext
  */
-char *url_adapt_ext(const char *src, size_t srclen, const char *name, const char *val, size_t *newlen)
+char *url_adapt_ext(const char *src, uint srclen, const char *name, const char *val, size_t *newlen)
 {
 	char buf[1024];
